requirements.txt
astroid==1.6.4
autopep8==1.3.5
backcall==0.1.0
bleach==2.1.3
blessings==1.6.1
bpython==0.17.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
curtsies==0.3.0
cx-Oracle==6.3.1
decorator==4.3.0
entrypoints==0.2.3
Flask==1.0.2
greenlet==0.4.13
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.4.0
Pygments==2.2.0
pylint==1.9.1
python-dateutil==2.7.3
python-decouple==3.1
python-dotenv==0.8.2
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.7
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
transitions==0.6.7
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
